Create function prepare treemap data.
