+++ src/ui/pages/chatgpt-message-dashboard.tsx
@@ -1,8 +1,8 @@
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
 import { HeatmapVisual } from "@/components/visualisations/heatmap-visual";
-import { Treemap, type Tree } from "@/components/visualisations/treemap";
-import { prepareHeatmapData } from "@/utils";
+import { Treemap} from "@/components/visualisations/treemap";
+import { prepareHeatmapData, prepareTreeMapData } from "@/utils";
 import { useEffect, useState } from "react";
 
 
@@ -47,38 +47,23 @@ export default function ChatGptMessageDashboard() {
     setDialogOpen(false)
   }
 
+  const dataWithValueCount = data.map(row => {
+    row.value = 1
+    return row
+  })
   const heatmapData = prepareHeatmapData(
-    data.map(row => {
-      row.value = 1
-      return row
-    }),
+    dataWithValueCount,
     "datetime",
     "value",
     "messages"
   )
 
-  const groupedDataMap = data.reduce((dataMap, row) => {
-    const role = row["role"]
-    if (!role) return dataMap
-    if (!dataMap[role]) {
-      dataMap[role] = 0
-    }
-    dataMap[role] += 1
-    return dataMap
-  }, {})
-  const flattenedGroupedData = Object.entries(groupedDataMap).map(([role, value])=> {
-    return {
-      type: "leaf",
-      name: role,
-      value: value as number
-    }
-  })
-  const treeData: Tree = {
-    type: "node",
-    name: "boss",
-    value: 0,
-    children: flattenedGroupedData.sort((a, b) => b.value - a.value) as Tree[]
-  }
+  const treemapData = prepareTreeMapData(
+    dataWithValueCount,
+    "role",
+    "value"
+  )
+
 
   return (
     <div className=' w-full h-fit p-3 flex flex-col gap-3'>
@@ -98,7 +83,6 @@ export default function ChatGptMessageDashboard() {
             <DialogContent>
               <DialogHeader>
                 <DialogTitle>Extract chatgpt data</DialogTitle>
-
                 <DialogDescription>
                   Select the chatgpt zip file you would like to process.
                 </DialogDescription>
@@ -118,24 +102,20 @@ export default function ChatGptMessageDashboard() {
                   Extract chatgpt message data
                 </Button>
               }
-
-
             </DialogContent>
           </Dialog>
-
         </div>
       </div>
 
       <div className='p-2 outline rounded-xl overflow-scroll h-50'>
-        <HeatmapVisual data={heatmapData} range={[0, 100]}/>
-      </div>
-      <div className='p-2 outline rounded-xl overflow-scroll h-50'>
-        <Treemap data={treeData} />
+        <HeatmapVisual data={heatmapData} range={[0, 100]} />
       </div>
 
+      <div className='p-2 outline rounded-xl overflow-scroll h-50'>
+        <Treemap data={treemapData} />
+      </div>
 
       <div className="font-light font-mono text-sm flex-1 wrap-break-word w-full bg-accent p-2 rounded-md">
-
         {JSON.stringify(data.slice(0, 10))}
       </div>
     </div>
