+++ src/ui/pages/screentime-dashboard.tsx
@@ -2,7 +2,7 @@ import { Button } from "@/components/ui/button";
 import { HeatmapVisual } from "@/components/visualisations/heatmap-visual";
 import MonthlyBarChart from "@/components/visualisations/montthly-barchart";
 import { Treemap, type Tree } from "@/components/visualisations/treemap";
-import { constructDurationString, prepareHeatmapData } from "@/utils";
+import { constructDurationString, prepareHeatmapData, prepareTreeMapData } from "@/utils";
 import * as d3 from "d3";
 import { useEffect, useState } from "react";
 
@@ -32,26 +32,12 @@ export default function ScreenTimeDashboard() {
   }, [])
 
 
-  const groupedByUsageMap = d3.rollup(
+  const treemapData = prepareTreeMapData(
     data,
-    v => d3.sum(v, d => d["usage"]),
-    d => d["app"]
+    "app",
+    "usage"
   )
-  const flattenedList = Array.from(groupedByUsageMap).map(([app, usage]) => {
-    return { app, usage }
-  })
-  const treeData: Tree = {
-    type: "node",
-    name: "boss",
-    value: 0,
-    children: flattenedList.sort((a, b) => b.usage - a.usage).map(row => {
-      return {
-        type: "leaf",
-        name: row["app"],
-        value: row["usage"]
-      }
-    })
-  }
+
 
   const groupedByMonthMap = d3.rollup(
     data,
@@ -85,7 +71,6 @@ export default function ScreenTimeDashboard() {
     row.label = constructDurationString(row.value)
     return row
   })
-  console.log(heatmapData)
 
   return (
     <div className=' w-full h-fit p-3 flex flex-col gap-3'>
@@ -107,7 +92,7 @@ export default function ScreenTimeDashboard() {
       </div>
 
       <div className='p-2 outline rounded-xl overflow-scroll flex justify-center h-80'>
-        <Treemap data={treeData}  />
+        <Treemap data={treemapData}  />
       </div>
 
       <div className='p-2 outline rounded-xl overflow-scroll flex justify-center'>
