+++ src/ui/utils.ts
@@ -1,4 +1,5 @@
 import type { InputHeatmapData } from "./components/visualisations/heatmap-visual";
+import type { Tree } from "./components/visualisations/treemap";
 
 
 
@@ -13,6 +14,13 @@ interface DateMapDataType {
 }
 
 
+/**
+ * so simple in sql so hard in ts 😮‍💨
+ * 
+ * SELECT dateTimeCol, SUM(valueCol), CONCAT(SUM(valueCol), " ${units}"
+ * FROM DATA
+ * GROUP BY date(dateTimeCol);
+ */
 export function prepareHeatmapData(
   fullData: Record<string, string | number>[],
   dateTimeCol: string,
@@ -46,4 +54,42 @@ export function constructDurationString(timeInSeconds: number): string {
   const hours = Math.floor(timeInSeconds / 3600)
   const minutes = Math.floor((timeInSeconds % 3600) / 60)
   return `${hours}h ${minutes}m`
+}
+
+
+/**
+ * SELECT categoryCol, SUM(valueCol) FROM data 
+ * GROUP BY categoryCol
+ */
+export function prepareTreeMapData(
+  fullData: Record<string, string|number>[],
+  categoryCol: string,
+  valueCol: string,
+) {
+  // const groupedData = Object.groupBy only in es2024 rip
+   const groupedDataMap = fullData.reduce((dataMap: {[category: string]: number}, row) => {
+      const category = row[categoryCol]
+      const value = row[valueCol] as number
+      if (!category || !value) return dataMap
+      if (!dataMap[category]) {
+        dataMap[category] = 0 as number
+      }
+      dataMap[category] += value
+      return dataMap
+    }, {} as {[category: string]: number})
+
+    const flattenedGroupedData = Object.entries(groupedDataMap).map(([role, value])=> {
+      return {
+        type: "leaf",
+        name: role,
+        value: value as number
+      }
+    })
+    const treeData: Tree = {
+      type: "node",
+      name: "no data to display!",
+      value: 0,
+      children: flattenedGroupedData.sort((a, b) => b.value - a.value) as Tree[]
+    }
+    return treeData
 }
